Refactoring...

+++ scripts/get_jmeter_reports.sh
@@ -30,9 +30,21 @@ shift "$((OPTIND - 1))"
 echo "Creating jmeter instance"
 gcloud compute instances create jmeter-instance --zone=${zone}
 
+if [ $? -ne 0 ]
+then
+  echo "Error getting auth...exiting."
+  exit $?
+fi
+
 echo "Attaching disks"
 gcloud compute instances attach-disk jmeter-instance --disk=jmeter-sync --zone=${zone}
 gcloud compute instances attach-disk jmeter-instance --disk=jmeter-async --zone=${zone}
+
+if [ $? -ne 0 ]
+then
+  echo "Error attaching disks...exiting."
+  exit $?
+fi
 
 echo "Compacting report files"
 gcloud compute ssh --zone=${zone} jmeter-instance -- '\
@@ -44,12 +56,36 @@ sudo tar -czf sync.tar.gz report/ results.csv && \
 cd /mnt/async/async && \
 sudo tar -czf async.tar.gz report/ results.csv'
 
+if [ $? -ne 0 ]
+then
+  echo "Error compacting files...exiting."
+  exit $?
+fi
+
 echo "Downling reports..."
 gcloud compute scp jmeter-instance:/mnt/sync/sync/sync.tar.gz ${output_dir}
 gcloud compute scp jmeter-instance:/mnt/async/async/async.tar.gz ${output_dir}
 
+if [ $? -ne 0 ]
+then
+  echo "Error downloading files...exiting."
+  exit $?
+fi
+
 echo "Removing report files from the server"
 gcloud compute ssh --zone=${zone} jmeter-instance -- 'sudo rm -rf /mnt/sync/sync /mnt/async/async'
 
+if [ $? -ne 0 ]
+then
+  echo "Error removing files...exiting."
+  exit $?
+fi
+
 echo "Deleting instance"
 gcloud compute instances delete jmeter-instance --quiet
+
+if [ $? -ne 0 ]
+then
+  echo "Error removing instance...exiting."
+  exit $?
+fi
